refactor: organize components for features
