+++ apps/client/src/lib/atoms/hello-atom.ts
@@ -1,0 +1,15 @@
+import { FetchHttpClient, HttpApiClient } from "@effect/platform";
+import { Api } from "@repo/domain";
+import { Effect } from "effect";
+import { runtime } from "../atom";
+
+const SERVER_URL = import.meta.env.VITE_SERVER_URL || "http://localhost:9000";
+
+export const helloAtom = runtime.fn(() =>
+  Effect.gen(function* () {
+    const client = yield* HttpApiClient.make(Api, {
+      baseUrl: SERVER_URL,
+    });
+    return yield* client.hello.get();
+  }).pipe(Effect.provide(FetchHttpClient.layer)),
+);
